Include D3SOX's Vencord userplugins
