--- src/excluded-plugins/D3SOX-userplugins/defaultStatusForever/README.md
@@ -1,5 +1,0 @@
-# DefaultStatusForever
-
-Makes the custom status screen default to "Don't Clear" as the duration
-
-![The custom status screen, with "Don't Clear" selected and as the first option](https://github.com/Vendicated/Vencord/assets/44179559/e8cf0ca1-589c-43ef-9a6d-4fa3ed03a1f9)
