+++ src/plusplugins/betterActivities/README.md
