+++ src/plusplugins/betterActivities/components/Caret.tsx
