+++ src/plusplugins/betterActivities/components/SpotifyIcon.tsx
