+++ src/plusplugins/betterActivities/styles.css
