+++ src/plusplugins/mediaPlaybackSpeed/components/SpeedIcon.tsx
