+++ src/plusplugins/silentTyping/README.md
