+++ src/plusplugins/silentTyping/index.tsx
