+++ src/plusplugins/simplifiedProfileNotes/index.tsx
